requirements.txt
kafka-python==2.0.2
loguru==0.5.3
psycopg2-binary==2.9.1
requests==2.25.1
pytest==6.2.5
responses==0.15.0

